added a makefile to generate a bourne compatable script
the bourne shell does not support $(()) so expr has to be used. This however slows everything down significantly and renders the game unplayable.
the main script uses $(()) in such a way that it can easily be translated to `expr ` syntax with a regex. The makefile just applies this regex
and generates a script that can run in the heirloom bourne shell

also fixed bug where quick drop wouldn't fasten the piece to the ground